pyproject.toml
[tool.poetry]
name = "tweetcrafter"
version = "0.1.0"
description = ""
authors = ["Venelin Valkov <venelin@curiousily.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
llamacloud-client = "^0.2.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
groq = "^0.4.0"
sambanova = "^0.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.5"
jupyterlab = "^4.2.1"
pandas = "^2.2.2"

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
